[mir] Mov instructions are no longer generated in MIR
Mov instructions are now only generated during SSA destruction, and are not present during all creation and optimization phases for MIR. Mov instructions were never necessary because MIR is in SSA form, so all uses of a copy can be replaced with an original. This also means that Phi instructions now must allow all values as their arguments. This commit mainly modifies SSA construction and constant folding to account for the lack of Phi nodes.

Note that this introduces a bug in SSA destruction, where sequentializing parallel copies when converting Phis to Movs clobbers a variable that is needed later. Fix coming soon.